requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
biopython==1.79
biothings-client==0.2.6
cached-property==1.5.2
celery==5.1.2
celery-progress==0.1.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.7
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-rest-framework==0.1.0
django-simple-history==3.0.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
idna==3.2
importlib-metadata==4.8.1
kombu==5.1.0
mygene==3.2.2
numpy==1.19.5
pandas==1.1.5
plac==1.3.3
prompt-toolkit==3.0.20
PyMySQL==1.0.2
python-crontab==2.5.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tqdm==4.62.2
typing-extensions==3.10.0.2
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
zipp==3.5.0
